# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: TikTok跨境电商全自动运营系统 API
  description: |
    TikTok跨境电商全自动运营系统MVP版本API接口契约。
    覆盖核心业务模块：认证、商品、订单、定价、上架、履约、客服、财务。
  version: "1.0.0"
  contact:
    name: Dev Team
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
security:
  - BearerAuth: []
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: 商品列表
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索关键词
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, inactive, listed, delisted]
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, name, sale_price, score]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
    post:
      tags: [Products]
      summary: 创建商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{product_id}:
    get:
      tags: [Products]
      summary: 商品详情
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: 商品详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Products]
      summary: 更新商品
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ProductIdParam:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        cost_price:
          type: number
          format: decimal
        sale_price:
          type: number
          format: decimal
        status:
          type: string
          enum: [draft, active, inactive, listed, delisted]
        score:
          type: number
          description: 选品评分(1-10)
        tiktok_product_id:
          type: string
          nullable: true
        source_url:
          type: string
          description: 1688商品链接
        skus:
          type: array
          items:
            $ref: "#/components/schemas/ProductSKU"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductSKU:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku_code:
          type: string
        specs:
          type: string
          description: JSON格式的规格键值对
        cost_price:
          type: number
          format: decimal
        sale_price:
          type: number
          format: decimal
        stock:
          type: integer
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        image_type:
          type: string
          enum: [main, detail, ai_generated]
        sort_order:
          type: integer
    ProductCreate:
      type: object
      required: [name, category_id]
      properties:
        name:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        source_url:
          type: string
        skus:
          type: array
          items:
            $ref: "#/components/schemas/ProductSKU"
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        sale_price:
          type: number
          format: decimal
        status:
          type: string
...
    PaginatedProducts:
      allOf:
        - $ref: "#/components/schemas/PaginatedBase"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Product"
...
    PaginatedBase:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...